Add files via upload
The internet is constantly growing and changing. A large number of  industries are use Hadoop to analyse their data,  whether it be a college, media company, art gallery , music label, and so on. As global internet traffic grows so does the amount of data. This became a problem for many industries and Hadoop was one of the first companies to offer a solution.